% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_immune.R
\name{immune_ode}
\alias{immune_ode}
\title{Within-host Immune ODE Model}
\usage{
immune_ode(time, theta, state, ...)
}
\arguments{
\item{time}{times to evaluate state variables at}

\item{theta}{named vector of parameters}

\item{state}{initial states}
}
\description{
write me
}
