% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito-Equilibrium.R
\name{jac_g}
\alias{jac_g}
\title{Numeric Jacobian of Linear Constraints Function for Mosquito Equilibrium}
\usage{
jac_g(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Calculate Jacobian matrix (1st order partial derivatives) of linear constraints
}
