% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito-Equilibrium.R
\name{make_obj_f}
\alias{make_obj_f}
\title{Objective Function for Mosquito Equilibrium}
\usage{
make_obj_f(theta, dt, NV, PL)
}
\arguments{
\item{theta}{named vector of parameters (see \code{\link{RACD_Parameters}})}

\item{dt}{size of time-step}

\item{NV}{total size of vector population}

\item{PL}{size of pupae population}
}
\description{
Calculate finite differences of discrete-time model
}
