\name{rACD-class}
\docType{class}
\alias{rACD-class}
\title{class: rACD Class}
\description{
Highest Level Virtual Package Class to which all other classes belong.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}
\section{Methods}{
No methods defined with class "rACD" in the signature.
}
\examples{
showClass("rACD")
}
\author{Alexios Ghalanos}
\keyword{classes}
