% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito-Equilibrium.R
\name{RACD_approx_equilibrium}
\alias{RACD_approx_equilibrium}
\title{Approximate (continuous-time) Equilibrium for Mosquito Model}
\usage{
RACD_approx_equilibrium(theta, lambdaV, IV)
}
\arguments{
\item{theta}{named vector of parameters (see \code{\link{RACD_Parameters}})}

\item{lambdaV}{equilibrium force of infection on mosquitos}

\item{IV}{number of infectious vectors at equilibrum}
}
\description{
Based on solutions of the mosquito model as a system of ODEs, return the values of EL, LL, K that
hold at equilibrium.
}
