% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mosquito-Equilibrium.R
\name{RACD_mosq_equilibrium}
\alias{RACD_mosq_equilibrium}
\title{Exact (discrete-time) Equilibrium for Mosquito Model}
\usage{
RACD_mosq_equilibrium(theta, dt, IV, lambdaV, sd = 0.25, nstart = 1000)
}
\arguments{
\item{theta}{named vector of parameters (see \code{\link{RACD_Parameters}})}

\item{dt}{size of time-step}

\item{IV}{number of infectious vectors at equilibrum}

\item{lambdaV}{equilibrium force of infection on mosquitos}

\item{sd}{size of box around approximate value is (x +- x*sd)}

\item{nstart}{number of starting points for optimization}
}
\description{
Using approximate solutions from \code{\link[RACD]{RACD_approx_equilibrium}} as starting point(s)
use constrained numerical optimization to find equilibrium solutions of (EL, LL, K)
from a given force of infection on mosquitoes and size of infected vector population,
which are obtainable from solving the human model at equilibrium.
Starting values for the optimzation are found by using Latin Hypercube Sampling
in a box around the approximate solutions.
}
