% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACD-Parameters.R
\name{RACD_Parameters}
\alias{RACD_Parameters}
\title{Generate Parameters for RACD Model}
\usage{
RACD_Parameters(epsilon0 = 5/365, fT = 0.4, dE = 12, dT = 5,
  dD = 5, dA = 200, dU = 110, dP = 25, cD = 0.068, cT = 0.32 *
  0.068, cU = 0.0062, gammaI = 1.82, rho = 0.85, a0 = 8,
  sigma2 = 1.67, d1 = 0.161, dID = 10 * 365, ID0 = 1.58,
  kappaD = 0.477, uD = 9.45, aD = 21.9, fD0 = 0.0071,
  gammaD = 4.81, alphaA = 0.757, alphaU = 0.186, b0 = 0.59,
  b1 = 0.5, dB = 10 * 365, IB0 = 43.9, kappaB = 2.16, uB = 7.2,
  phi0 = 0.792, phi1 = 0.00074, dC = 30 * 365, IC0 = 18,
  kappaC = 2.37, uC = 6.06, PM = 0.774, dM = 67.7, rW = 0.723,
  rP = 0.342, meanAge = 17.4, N = 200, meanNumPeoplePerHouse = 6.5,
  numHousesPerBreedingSite = 5, beta = 21.19, muEL = 0.034,
  muLL = 0.035, muPL = 0.25, durEL = 6.64, durLL = 3.72,
  durPL = 0.64, durEV = 10, gamma = 13.25, tau1 = 0.68,
  tau2 = 2.32, muV = 1/7.6, Q0 = 0.92, phiB = 0.89, phiI = 0.97,
  rITN = 0.56, sITN = 0.03, rIRS = 0.6, sIRS = 0)
}
\arguments{
\item{epsilon0}{Mean EIR for adults (per day)}

\item{fT}{Proportion of clinical disease cases successfully treated}

\item{dE}{Duration of latent period (days)}

\item{dT}{Duration of treated clinical disease (days)}

\item{dD}{Duration of untreated clinical disease (days)}

\item{dA}{Duration of patent infection (days)}

\item{dU}{Duration of sub-patent infection (days) (fitted)}

\item{dP}{Duration of prophylactic protection following treatment (days)}

\item{cD}{Infectiousness with untreated disease & no immunity (fitted)}

\item{cT}{Infectiousness after treatment}

\item{cU}{Infectiousness with sub-patent infection (fitted)}

\item{gammaI}{Relates infectiousness to probability of detection (fitted)}

\item{rho}{Age-dependent biting parameter}

\item{a0}{Age-dependent biting parameter (years)}

\item{sigma2}{Variance of log of heterogeneity in biting rates}

\item{d1}{Probability of detection with maximum immunity (fitted)}

\item{dID}{Inverse of decay rate (days)}

\item{ID0}{Immunity scale parameter (fitted)}

\item{kappaD}{Immunity shape parameter (fitted)}

\item{uD}{Duration in which immunity is not boosted (days) (fitted)}

\item{aD}{Scale parameter relating age to immunity (years) (fitted)}

\item{fD0}{Parameter relating age to immunity (fitted)}

\item{gammaD}{Shape parameter relating age to immunity (fitted)}

\item{alphaA}{PCR prevalence parameter (fitted)}

\item{alphaU}{PCR prevalence parameter (fitted)}

\item{b0}{Probabiliy with no immunity (fitted)}

\item{b1}{Maximum relative reduction}

\item{dB}{Inverse of decay rate (days)}

\item{IB0}{Scale parameter (fitted)}

\item{kappaB}{Shape parameter (fitted)}

\item{uB}{Duration in which immunity is not boosted (days) (fitted)}

\item{phi0}{Probability with no immunity}

\item{phi1}{Maximum relative reduction}

\item{dC}{Inverse decay rate (days)}

\item{IC0}{Scale parameter}

\item{kappaC}{Shape parameter}

\item{uC}{Duration in which immunity is not boosted (days)}

\item{PM}{New-born immunity relative to mother's immunity}

\item{dM}{Inverse decay rate of maternal immunity (days)}

\item{rW}{Weekly active case detection}

\item{rP}{Weekly passive case detection}

\item{meanAge}{Mean age in Tanzania (males and females, years)}

\item{N}{Village population size}

\item{meanNumPeoplePerHouse}{Mean number of people per house (from Misungu data set)}

\item{numHousesPerBreedingSite}{Number of houses per breeding site}
}
\description{
Return a named vector of parameters.
}
