% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RACD-Setup.R
\name{RACD_Setup}
\alias{RACD_Setup}
\title{Initialize State for RACD Model}
\usage{
RACD_Setup(theta)
}
\arguments{
\item{theta}{named vector of parameters (see \code{\link{RACD_Parameters}})}
}
\description{
DEBUG VERSION Initialize initial conditions for the RACD simulation model
}
\examples{
\dontrun{
library(RACD)
library(tidyverse)
library(reshape2)
theta <- RACD_Parameters(N=1e3)
init <- RACD_Setup(theta)
out <- RACD_Simulation(365*10,theta,init$humans,init$houses,42)
out <- as.data.frame(out)
out$time <- as.integer(rownames(out))
out <- melt(out,id.vars = "time")
ggplot(data = out,aes(x=time,y=value,color=variable)) + geom_line()
}

}
