% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_house}
\alias{test_house}
\title{Unit Test: house}
\usage{
test_house(theta, seed)
}
\description{
test putting a human in a house and calling some basic actions
}
\examples{
test_house(theta = RACDaux::RACD_Parameters(),seed = 1)
}
