% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_human_parameters}
\alias{test_human_parameters}
\title{Unit Test: test_human}
\usage{
test_human_parameters(theta)
}
\arguments{
\item{theta}{output of \code{\link[RACDaux]{RACD_Parameters}}}
}
\description{
Unit Test: test_human
}
