% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_list}
\alias{test_list}
\title{Unit Test: nested lists}
\usage{
test_list(list)
}
\description{
test iteration through nested list
}
\examples{
theta = RACDaux::RACD_Parameters()
init = RACDaux::RACD_Setup(theta)
test_list(init$humans)
}
