% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_parameters}
\alias{test_parameters}
\title{Unit Test: RACD_Parameters}
\usage{
test_parameters(theta)
}
\arguments{
\item{theta}{output of \code{\link[RACDaux]{RACD_Parameters}}}
}
\description{
Unit Test: RACD_Parameters
}
