% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_village}
\alias{test_village}
\title{Unit Test: make a village}
\usage{
test_village(human_par, house_par)
}
\description{
create a village
}
\examples{
theta = RACDaux::RACD_Parameters()
init = RACDaux::RACD_Setup(theta)
human_par = init$humans
house_par = init$houses
}
