\name{ACDmultispec-class}
\docType{class}
\alias{show,ACDmultispec-method}
\alias{ACDmultispec-class}
\title{class: Univariate ACD Multiple Specification Class}
\description{Class for the univariate ACD Multiple specification.}
\section{Objects from the Class}{A virtual Class: No objects may be created from it.}

\author{Alexios Ghalanos}
\seealso{
Classes \code{\linkS4class{ACDmultiforecast}}, \code{\linkS4class{ACDmultifit}} and \code{\linkS4class{ACDmultifilter}}.
}
\keyword{classes}