% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode_immune.R
\name{infection_initial}
\alias{infection_initial}
\title{Within-host Infection State ODE Model}
\usage{
infection_initial(a, zeta, psi, epsilon0 = (0.01369863 * 365), fT = 0.4,
  dE = (12/365), dT = (5/365), dD = (5/365), dA = (200/365),
  dU = (110/365), dP = (25/365), rho = 0.85, a0 = 8, b0 = 0.59,
  b1 = 0.5, dB = (3650/365), IB0 = 43.9, kappaB = 2.16,
  durB = (3650/365), uB = (7.2/365), durD = (3650/365), uD = (9.45/365),
  durC = (10950/365), uC = (6.06/365), phi0 = 0.792, phi1 = 0.00074,
  dC = (10950/365), IC0 = 18, kappaC = 2.37, PM = 0.774,
  dM = (67.7/365), ...)
}
\arguments{
\item{a}{age of individual}

\item{zeta}{individual biting heterogeneity}

\item{psi}{relative biting rate (used to calculate psi(a))}

\item{epsilon0}{mean EIR}

\item{fT}{proportion of clinical disease cases successfully treated}

\item{dE}{Duration of latent period (years)}

\item{dT}{Duration of treated clinical disease (years)}

\item{dD}{Duration of untreated clinical disease (years)}

\item{dA}{Duration of patent infection (years)}

\item{dU}{Duration of sub-patent infection (years) (fitted)}

\item{dP}{Duration of prophylactic protection following treatment (years)}

\item{rho}{age-dependent biting heterogeneity (used to calculate psi(a))}

\item{a0}{age-dependent biting heterogeneity (used to calculate psi(a))}

\item{b0}{probability of infection with no immunity}

\item{b1}{maximum relative reduction in immunity}

\item{dB}{Inverse of decay rate (years) of immunity reducing probability of infection}

\item{IB0}{scale parameter for probability of infection}

\item{kappaB}{shape parameter for probability of infection}

\item{durB}{durB 1/decay rate of immunity reducing probability of infection}

\item{uB}{duration in which immunity is not boosted}

\item{durD}{1/decay rate of immunity reducing probability of detection}

\item{uD}{duration in which immunity is not boosted}

\item{durC}{1/decay rate of immunity reducing probabiltiy of clinical disease}

\item{uC}{duration in which immunity is not boosted}

\item{phi0}{probability of clinical disease with no immunity}

\item{phi1}{maximum relative reduction of clinical disease for immunity reducing probability of clinical disease}

\item{dC}{1/decay rate of immunity reducing probability of clinical disease}

\item{IC0}{Scale parameter of immunity reducing probability of clinical disease}

\item{kappaC}{Shape parameter of immunity reducing probability of clinical disease}

\item{PM}{New-born immunity relative to mother's immunity}

\item{dM}{Inverse decay rate of maternal immunity (years)}

\item{...}{additional parameters passed to numerical integrator \code{\link[deSolve]{ode}}}
}
\description{
ODE model for calculating the probability that an individual is in each state given their age and EIR heterogeneity attributes.
}
