% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RACD_Simulation}
\alias{RACD_Simulation}
\title{RACD Simulation Model}
\usage{
RACD_Simulation(tMax, theta, human, house, seed)
}
\arguments{
\item{tMax}{length of simulation in days}

\item{theta}{named list of parameters (see \code{\link[RACDaux]{RACD_Parameters}} for details)}

\item{human}{list of human parameters (see \code{\link[RACDaux]{RACD_Setup}} for details)}

\item{house}{list of house parameters (see \code{\link[RACDaux]{RACD_Setup}} for details)}

\item{seed}{seed for prng class}
}
\description{
Run the simulation
}
\examples{
theta = RACDaux::RACD_Parameters()
init = RACDaux::RACD_Setup(theta)
\dontrun{RACD_Simulation(365,theta,init$humans,init$houses,123)}

}
