% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RACD_Simulation}
\alias{RACD_Simulation}
\title{RACD Simulation Model}
\usage{
RACD_Simulation(tMax, theta, human, mosy, house)
}
\arguments{
\item{tMax}{length of simulation in days}

\item{theta}{named list of parameters (see \code{\link{RACD_Parameters}} for details)}

\item{human}{list of human parameters (see \code{\link{RACD_Setup}} for details)}

\item{house}{list of house parameters (see \code{\link{RACD_Setup}} for details)}

\item{seed}{seed for prng class}

\item{outfile}{path to .csv file for logging state transition events}
}
\description{
Run the simulation
}
\examples{
\dontrun{
library(RACD)
library(tidyverse)
library(spatstat)
xy_h <- rpoispp(lambda = 100,win = owin(c(0,1),c(0,1)))
xy_b <- rpoispp(lambda = 100,win = owin(c(0,1),c(0,1)))
theta <- RACD_Parameters()
init <- RACD_Setup(as.matrix.ppx(xy_h),as.matrix.ppx(xy_b),theta)
outfile = "/Users/slwu89/Desktop/log_trans.csv"
RACD_Simulation(365,theta,init$humans,init$houses,123,outfile)
state = RACD_StateVector(outfile)
state \%>\% as.tibble \%>\% gather(state,value,-time) \%>\% ggplot(aes(x=time,y=value,color=state)) + geom_line() + theme_bw()
}

}
